fix css issue

+++ apps/browser/src/popup/accounts/lock.component.ts
@@ -105,13 +105,13 @@ export class LockComponent extends BaseLockComponent {
           const offsetTop = rect2.top - rect1.top;
           const offsetLeft = rect2.left - rect1.left;
           pass2.style.top =
-            parseInt(
+            parseFloat(
               window.getComputedStyle(pass2, null).getPropertyValue("top").replace("px", "")
             ) -
             offsetTop +
             "px";
           pass2.style.left =
-            parseInt(
+            parseFloat(
               window.getComputedStyle(pass2, null).getPropertyValue("left").replace("px", "")
             ) -
             offsetLeft +
